fix: wait until 10 successes to consider a runner healthy
when a runner goes live we had a few cases where the fleet supervisor
would successfully reach the /health endpoint and mark it as running but
then a script execution would fail trying to reach the runner.
My hypothesis is that Render network config can take a bit of time to
propagate and even if a runner can be reached from one jobs instance
(the one running fleet supervisor) there is not guarantuee it can also
be reached immediately by all jobs instances.
This commit requires consecutive successes when making sure the runner
is healthy before it can be registered. It is making the registration
slower but it is a one-off thing.