fix: pause schedules when disabling integration
when disabling an integration we currently only disable the sync config
but we don't pause the schedules, which means we have schedules in
STARTED state which are linked to disabled config.
The orchestrator is scheduling tasks for this schedules, even though
the tasks fails quickly because the sync config is disable.
Once this fix is merged, I will go over the existing schedules and
paused every one of them that is associated to a disabled sync config