fix: pause schedules when disabling integration (#3519)
when disabling an integration we currently only disable the sync config
but we don't pause the schedules, which means we have schedules in
STARTED state which are linked to disabled configs. The orchestrator is
scheduling tasks for this schedules, even though the tasks fails quickly
because the sync config is disable.
Once this fix is merged, I will go over the existing schedules and
paused every one of them that is associated to a disabled sync config.

How to test:
1. Disable/enable integrations endpoints
2. check in db that the schedules state changes to PAUSED/STARTED