feat: introduce sync variant
Adding the 'variant' column to the syncs table with a default 'base'value
and add variant parameters to sync related function so we can pass a
different value later.
This commit doesn't change any behavior. There is still now way to
create a variant and cardinality of sync_config:sync is still 1:1
Next steps:
1. modify existing public function/api (trigger, pause, status, ...) to accept a sync name + variant
2. expose variant into the script
3. modify runner-sdk to save variant records and modify getRecords to handle variants
4. add new functions/api to be able to create a variant
5. update docs