feat: introduce sync variant (#3521)
Adding the 'variant' column to the syncs table with a default
'base'value and add variant parameters to sync related function so we
can pass a different value later.
This commit doesn't change any behavior. There is still no way to create
a variant and cardinality of sync_config:sync is still 1:1

Next steps:
- modify existing public function/api (trigger, pause, status, ...) to
accept a sync name + variant
- expose variant into the script
- modify runner-sdk to save variant records and modify getRecords to
handle variants
- add new functions/api to be able to create a variant
- update docs