feat(sync variant): add sync variant to schedule and nangoProps (#3522)
Passing sync variant as part of schedule properties (adding support for
variant to jobs and orchestrator along the way) so it can be exposed via
NangoProps. For now we are doing nothing with it.

How to test:
- check that existing syncs can run successfully
- deploy a new sync or enable a pre-built one that you have never
enabled. After running it check in the db that schedules and tasks now
have a syncVariant field in their payload