fix(connectionRefresh): force salesforce refresh on error, add metrics (#3530)
## Changes

- Force salesforce refresh on introspection error
When introspection failed, the method returned `false` which meant no
need to refresh, except that an outdated `access_token` would throw a
400. Not sure if it can spiral into a negative loop, but the opposite
would miss some refresh and not log anything.

- Add metrics
We have no volume of refresh, except when the cron jobs kicks in which
prevents us on monitoring the global volume of error.

- Differentiate connection test success and not ran