chore: Integrations changelog (#3526)
<!-- Describe the problem and your solution --> 
This dumps markdown for Bastien to use for reporting our changelog for
use cases. It reads flows.yaml over time. It seems like it's not very
good before November of last year due to changes to the flows format,
but it should be useful going forward.

Runnable with:

```
npm run changelog:integrations
```

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

---------

Co-authored-by: Samuel Bodin <1637651+bodinsamuel@users.noreply.github.com>