fix(es): add indexed retry object for http logs (#3543)
## Changes

Contributes to
https://linear.app/nango/issue/NAN-2736/new-retry-strategy-for-proxy
Contributes to
https://linear.app/nango/issue/NAN-2683/log-retries-in-a-way-that-it-can-be-use-to-graph

- Prepare the **retry** field
It will contain information related to each http log, whether they are
part of a retry, the attempt number, and the previous wait time. It's
all based on theoretical test so hopefully I nailed it.
Note it's the first undefined field I create, I realized nullable field
are not ideal in terms of storage cost, so I had to change also some
types