feat(sdk): add uncontrolledFetch (#3564)
## Changes

- Add `uncontrolledFetch` 
To be able to just fetch some resources that are unrelated to connection
or provider. It's using `undici` because if we want to get rid of axios
it's probably better to not add more here.
Naming is debatable :D 

- Remove some leftover 


## Tests

```ts
import type { NangoSync } from "../../models";

export default async function fetchData(nango: NangoSync): Promise<void> {

  const res = await nango.uncontrolledFetch({
    url: new URL("https://httpstatuses.maor.io/500"),
  });

  await nango.log("Res", { body: await res.json() });
}
```