fix(proxy): remove token from payload, rewrite headers building (#3557)
## Changes

- Remove `token` from payload
I'm not sure why it was used, probably a legacy thing. It wasn't typed
correctly and was unnecessary since we carry the credentials in the
`connection`. It removes a big chunk of code.

- Rewrite header building 
It was confusing with a default case that was almost always wrong, that
got override by custom headers.