fix(proxy): refresh connection between retries  (#3566)
## Changes

Fixes
https://linear.app/nango/issue/NAN-2745/retry-strategy-should-call-getconnection-before-each-try

> credentials are expired between calls


- ProxyRequest now accepts a `getConnection` param called on every
iteration
The main idea behind this PR is to re-inject the connection after TTL
expires.
I had to rewrite part of the code to be able to dynamically build axios
config (url + headers).


- Type: Split ProxyConfig and Connection
Passing a connection to this type was not useful and was kind of weird
with the new behavior. Now we pass a base proxyconfig that tells what we
want to query and we recompile the axios config based on that.


## Tests

- Call the proxy with the retry header
```
curl --request GET \
  --url http://localhost:3003/proxy/500 \
  --header 'Authorization: REDACTED' \
  --header 'Base-Url-Override: https://httpstatuses.maor.io' \
  --header 'Connection-Id: 025d9a38-f31c-4a64-b034-57a425761a90' \
  --header 'Provider-Config-Key: unauthenticated' \
  --header 'Retries: 10'
```

- Same for local sync
```
import type { NangoSync } from "../../models";

export default async function fetchData(nango: NangoSync): Promise<void> {
  await nango.log("started");
  await nango.proxy({
    baseUrlOverride: "https://httpstatuses.maor.io",
    endpoint: "/500",
    method: "GET",
    retries: 10
  });
}
```

**Very long retry with a periodic refresh**
![Screenshot 2025-02-25 at 12 13
08](https://github.com/user-attachments/assets/67d18e45-d296-4fcd-85b9-0e91fc9607d2)