fix(otlp): manual span (#3647)
## Changes

Contributes to
https://linear.app/nango/issue/NAN-2803/investigate-otlp-export

- Manually declare OTLP span
We need to dissociate span creation from logCtx, because getting the
operation entirely from ES is not a viable solution. This fix only
changes Span to be declared manually but does not fix the main issue.
Just wanted to be sure it was doable and limit the scope of this PR.

- Fix nangoProps type
- Fix OTLP export broken by Sentry (skipOtel)