chore: add GET /records metrics (#3677)
We want to be able to know how much records/data is being read for each
account from the records db.
Adding metrics on GET /records endpoint:
- number of records fetched
- size of records fetched (in order to avoid calculating the size, we
use the content-length of the response. It is not exactly the size of
the records since the response contains some metadata like cursors but
it is a good and stable approximation)

Those 2 metrics would be a good proxy for records db I/O