fix(connection): split refresh function (#3674)
## Changes 

- Move connection refresh method (and directly associated method) into
it's own file
connection.service.ts is hard to parse and maintain, since I need to
touch it a lot I figured it was time to do something. I tried to keep it
simple, I didn't move everything and tried to keep the code almost
similar.
There is a huge refactor pending this but not today


#### Notable changes:
- refreshOrTestCredentials now uses a switch case and call more function
for readability
- refresh and test functions now always return the new connection,
because we are going to modify more than just credentials, it was a
footgun before