fix(logs): compute index name from operation id (#3719)
## Changes

Fixes
https://linear.app/nango/issue/NAN-2925/use-id-to-compute-index-name

- Compute index name from operation id
This finally gets rid of using Redis and es search to get back the
operation.
To limit the changes and still keep the library easy to use I had to
create a new sub class that still contains the operation, but as soon as
you get the operation back from another service it does not accept an
operation so it's not possible to break this part.
Because of this, I had to pass the `accountId` manually where needed, as
planned.

Hopefully, this should drastically improve perf, if not I do 100 push-up
🫡