chore(eslint): import order as warning (#3721)
## Changes

- Enable `import/order` as warning
The goal is to provide consistency in our codebase, improve readability,
and limit the noise when changing import.

This means all changed files will be automatically linted. Never too
sure if it's the best strategy, but it's sometimes better than having
one large commit that doesn't mean much.
The strategy is:
```
- side effect import `import ./tracer.js`
- built-in node package
- external npm package
- our own 
- local import 
- types
```
Import and object import are also sorted alphabetically


- Tested on a few file just for you to see