fix(logs): buffer context, log oauth2 http calls (#3765)
## Changes

Contributes to
https://linear.app/nango/issue/NAN-2975/investigate-failed-refresh

- Add log context buffer class
I got annoyed at buffering logs in an array and passing up the array. We
now have a buffer class that acts like a regular logCtx so they are used
without knowing if it's a buffer or not.

- Pass log buffer to oauth2 client
I was trying to debug something for a customer but we were not logging
oauth2 calls. simple-oauth2 is hiding a lot of stuff, it's really
annoying