fix(connection): decryption without a secret key would void credentials (#3769)
## Changes

User reported an issue, but couldn't reproduce it. Turns out it was a
bug with decryption when encryption_key is not set, which is quite rare
since we are recommending NANGO_ENCRYPTION_KEY to be set for a while now
(edit: we are not enforcing which might be a mistake)

- Fix decryption should return credentials even when encryption is not
enabled
I had copied this code but before there was an early return that would
prevent this code from being reached.
Added small unit test to prevent regression