chore: sync total active time
We want to track how long syncs are being active (aka: started)
In order to get an approximation, we are pushing a metric per account
for every sync being started. The value is equal to the frequency of the
sync. By summing up this metric we should then be able to know the total
active sync time for an account