fix: make Slack notifications aware of sync variants (#3771)
Slack notifications are currently not aware of sync variants, which
means that all sync execution for the base sync and its variant fall
into the same entry in the slack_notifications table. This could lead to
bugs where one variant fails triggering the push of a 'Failed sync'
notif followed by another variant that is successful, marking the
initial Slack notifications as resolved even though it is a different
variants.
Fixing the bug by appending the variant to the sync name if it is not
the base one