fix(auth): oauth2_cc support authorization_params (#3792)
## Changes

Fixes
https://linear.app/nango/issue/NAN-2996/support-authorization-params-in-oauth2-cc

- oauth2_cc support authorization_params
We were not forwarding the param like we are doing in oauth2. Also
saving them in connectionConfig (unfortunately I wasn't really aware
that it would be better in connectionConfig so I need to do it manually
now)

- Introduce `loggedFetch` 
I noticed none of the auth have clear trace in the system so I tried to
put some logging and realised it would be super repetitive. So it will
help replace all of them with the same code.

---
![Screenshot 2025-03-27 at 15 12
48](https://github.com/user-attachments/assets/0a3351d8-f45a-472e-a1af-7fa8ae4e35fe)
![Screenshot 2025-03-27 at 15 11
58](https://github.com/user-attachments/assets/0ae7fb8b-e299-4def-855c-7cd76879194b)