fix: do not block request if rate limit cannot be computed (#3806)
As experienced recently, a redis outage would cause all requests made to
Nango api and web to fail.
This commit changes the behavior to not block requests if rate limit
cannot be computed, which will allow requests to dashboard and api to
continue to be served even if redis isn't reachable
Caveat: if Redis is down we don't have any rate-limit protection but it
is better imho than to be fully down. wdyt?