chore(test): provide more seeder for unit test (#3823)
## Changes

- Provide more seeder for unit test
We are often using type casting to fit the type requirement but it's
easier to have a correct seed. I wanted to get rid of more but some are
in logs and webhooks which are not requiring `shared` folder. I guess
there is more split to do, but not sure which one, maybe a models
package

- Moved getTestConnection to be reusable