docs: When there is a connect ui document, link to that for the setup guide (#3751)
1. Added a `docs_connect` property to integrations that have a setup
guide but lacks the connect property
2. Implemented a pre-commit hook which automatically generates `docs`
and `docs_connect` properties for integrations which already have a
`connect.mdx` documentation file in them.

## How `2` works:
The hook works by retrieving all integrations in the
`docs-v2/integrations/all` folder that has a `connect.mdx` file. Next,
it performs a diff of integrations from the `providers.yaml` file which
already have either `docs` or `docs_connect` properties to determine
which needs to be updated. It then makes the updates and writes the
result back to the `providers.yaml` file.

### Known issues
If an integration in the `providers.yaml` file has an alias, the hook
will inadvertently copy properties from the aliased integration into the
said integration. This is because the `loadProvidersYaml` function which
loads providers also denormalizes aliases before it can be processed.
However, this doesn't affect the integration since it is an explicit
form of the `alias` feature of the `providers.yaml` file

Ticket link:
https://linear.app/nango/issue/EXT-556/when-there-is-a-connect-ui-document-link-to-that-for-the-setup-guide

---------

Co-authored-by: Samuel Bodin <1637651+bodinsamuel@users.noreply.github.com>