fix(plan): use plan for tracer (#3848)
## Changes

Fixes https://linear.app/nango/issue/NAN-3020/update-tracer-with-plan
Contributes to
https://linear.app/nango/issue/NAN-3019/store-ratelimit-configuration-in-plan

- Use plan for tracer
It's only partially used, in jobs we don't really need it for the
moment.

- Added new column to store api rate limit
I have a big question here, rate-limit-flexible does not allow points to
be changed dynamically. What we currently do is increase/decrease the
consumption based on some condition but it's not really precise. I have
looked for ways to do it natively but there is no way. What I have done
in the past is to create one class per customers, but not sure with our
scale it would fit.

```ts
const limiters = new Map();
function getLimiterForCustomer(customerId, points, duration = 60) {
  if (!limiters.has(customerId)) {
    const limiter = new RateLimiterMemory({ points, duration });
    limiters.set(customerId, limiter);
  }

  return limiters.get(customerId);
}
```
@TBonnin