feat: nango locks in scripts
Adding functions to the nango runner-sdk to manage locks.
```
await nango.tryAcquireLock({ key: "abc", ttlMs: 60_000 })
```
It makes it possible for customers to handle custom locking logic in
their script.
I decided to make the locking a runner-sdk implementation details and
be ok with a bit more duplication of code rather than having the
implementation as part of the runner-sdk abstract class (and then have
the state not fully self-contained). Locks state is imho at the
runner-level and not script-level.