feat: nango locks in scripts (#3863)
Adding functions to the nango runner-sdk to manage locks.
```
await nango.tryAcquireLock({ key: "abc", ttlMs: 60_000 })
```
It makes it possible for customers to handle custom locking logic in
their script, like skipping a webhook execution if one is already
running.
I decided to make the locking a runner-sdk implementation details and be
ok with a bit more duplication of code rather than having the
implementation as part of the runner-sdk abstract class (and then have
the state not fully self-contained). Locks state is imho at the
runner-level and not script-level.

Caveat: for now the locks are stored in memory (by the runner). During
deployment of runners, there might be 2 runners up at the same time and
potentially the same lock could be acquired twice. In a second steps we
will store the locks in a shared datastore to ensure 100% unique locks

# How to test
You can have 2 syncs that both tries to acquire the same lock and wait
to ensure overlap
```
    const lock = await nango.tryAcquireLock({
        key: 'MY_KEY',
        ttlMs: 60_000,
    });
    if (!lock) {
        await nango.log('could not acquire lock. skipping...');
        return
    }
    await nango.log('acquired lock');

    const waitSecs = 10;
    for (let i = 0; i < waitSecs; i += 1) {
        await new Promise((resolve) => setTimeout(resolve, 1000));
        await nango.log(`waiting for ${waitSecs - i} seconds`);
    }
```