fix(runner-sdk): no need for a mutex
Mutex is not necessary when acquiring/releasing lock. No race condition
can happen because of nodejs concurrency model.
I think I had a mutex because in my original implementation there were
an async operation between the get and set (making it possible for the
execution to yield) but that's not the case now