fix(ui): logs fix history, virtualization, data loading and caching (#3888)
## Changes

Fixes https://linear.app/nango/issue/NAN-2995/improve-log-navigation

Big changes regarding logs interface. And it's not done yet

- **Fix history navigation**
I coded something badly regarding query params. I knew at the time by
experience that when you need a two-way sync between page and query
params you are in trouble. I decided to rewrite this logic and found
`nuqs` which basically does that plus better parsing.

- **Add virtualization for operations list**
When loading a lot of rows in a UI you usually quickly hit a bottleneck
and/or crash, mostly because of DOM, not memory. Virtualization keeps
the DOM short and the rest in memory.
It now supports thousand of operations and loads them much faster

- **Fix data loading and caching**
When I did this page we only had SWR which doesn't support POST. I
immediately hated Vercel a little bit more but decided to bite the
bullet and recode my own data fetching. Building a UI is hard and it's
easy to get lost in `useEffect` and such. The code was definitely messy
and not always working.
So I decided to stop pretending SWR is good and introduce tanstack
react-query which is ultimately a better solution for data fetching in
every way.
I replaced my code with `useInfiniteQuery` which does 90% of what I
wanted, except one annoying thing about interval refresh refreshing
every page.
On top of that, we have better caching and un-caching

## 🧪 Tests

- Stress the UI


https://github.com/user-attachments/assets/c5d21aa8-c12e-4ec6-a928-fa04040b7cbe

- You can use autocannon to push more operations
```sh
npx autocannon -a 10000 http://localhost:3003/action/trigger \
-H"Authorization: REDAC" \
--header 'Connection-Id: 765b3a78-11ed-402b-8385-09708519ffad'\
--header 'Content-Type: application/json'\
--header 'Provider-Config-Key: unauthenticated' \
-m POST -b '{"action_name": "returnBase","input": { "foobar": "foobar"}}'