feat: delete environments (#3899)
<!-- Describe the problem and your solution --> 
## Description
This PR adds the ability to (soft) delete environments, from frontend to
backend.

### Frontend
A `Delete environment` button was added to the environment-settings
(doesn't show up for the prod environment).

![image](https://github.com/user-attachments/assets/6782a09f-8c47-491f-9833-8a78cc76a0ed)

A `DestructiveActionModal` component requires the environment name to be
typed in to confirm deletion.

![image](https://github.com/user-attachments/assets/9aecc8ae-dfd8-48bf-a3ce-c229270f24c6)


### Backend
The environment is soft-deleted along with:
- It's Provider/Integration configs (nango_config) - Ps: we need to
consolidate on a name for these
- Syncs - and schedules are deleted
- Sync Configs
- Connections

Hard deletion of old soft-deleted environments (and it's dependents)
will be added in a following PR.

<!-- Testing instructions (skip if just adding/editing providers) -->
## Testing
Please checkout to branch and test deleting environments in the UI.
Also, help me verify that anything that wastes computation or would
impact billing is stopped / soft-deleted by my changes.