fix(ui): disable cache when live (#3921)
## Changes

- Disable cache when live
When you have cache you have more problems. Tried to be smart, but it
wasn't always working, especially when discarding the cursor it would at
some point refresh the wrong page. Changed to disable cache when we are
live (top of the page or in a running script). And use manual refetch
for operations too