feat(persist): log unchanged records (#3929)
## Changes

Fixes https://linear.app/nango/issue/NAN-2916/log-unchanged-records

- Log unchanged records 
In #3496 I tried to do that, but also changing the webhook body, which
created some debate and potential breaking changes. Took a simpler
approach this time by just modifying the log. We need this for
transparency but also to be able to really find records by ID. I expect
this change to have a **bigger impact on log storage** but we'll see


(*don't mind the singular it's fixed*)
![Screenshot 2025-04-18 at 11 31
55](https://github.com/user-attachments/assets/60abec79-a61a-4245-997a-a67f3bacf4e3)
Inside the log msg
![Screenshot 2025-04-18 at 11 32
08](https://github.com/user-attachments/assets/e7395c5b-e4b6-47c3-930b-0ff0dfd8d649)