feat(logs): global search (#3906)
## Changes

Fixes https://linear.app/nango/issue/NAN-3039/top-level-log-search

- Logs global search

You will see some comments in the PR, but basically, the main challenge
is that it's a **two-step process**. First, we do a regular search on
operations with the filters (if any), and then we search inside the
`operationIds` that were found.

**Doing post-filtering** instead of pre-filtering saves a lot of
headaches for pagination and security, but it's not perfect.
Since the pagination is working (we have a valid cursor) but the limit
might be incorrect (we asked for 20 and got less) the UI will ask for
more due to the infinite scroll.
To prevent excessive HTTP calls, I have added **a manual mode**. As soon
as you type, we stop the live refresh and disable infinite scroll. It
will still try to load as much as it can when we possibly, but the
initial load is always manual.

I'm actually surprised the PR is that small, hopefully it's not bringing
too much bugs 🙏🏻



https://github.com/user-attachments/assets/a1362512-986f-4373-8255-3645144e938e