fix: upgrade dependencies (#3957)
## Changes

- Upgrade dev dependencies
Vite, vitest, type-fest, eslint-*, etc.

- Upgrade dependencies that uses Axios to fix CVE
axios, botbuilder, soap, posthog
It's not fixing all of them because mailgun is not upgrading and some
part of botbuilder are not using the same version

    
<!-- This is an auto-generated description by mrge. -->
---

## Summary by mrge
Upgraded dev and production dependencies across all packages to fix
security issues and keep libraries up to date.

- **Dependencies**
- Updated axios and related packages to address security
vulnerabilities.
  - Upgraded Vite, Vitest, type-fest, eslint, and other dev tools.

<!-- End of auto-generated description by mrge. -->