fix(persist): adapt wording for batchDelete, log missing error (#3958)
## Changes

- Adapt wording and lock name for batchDelete
- Notice we were not logging error which was confusing 
They only appeared in the operation payload which I never check


@TBonnin I found a "bug" in the batchDelete that took me a while to
understand. If you batchDelete a record that never existed in the first
place, it will create it (which is debatable), but because it actually
creates it, the computed stats are also wrong. Didn't find an easy
solution


    
<!-- This is an auto-generated description by mrge. -->
---

## Summary by mrge
Updated log messages for batchDelete to use clearer wording and lock
names, and added missing error logging for better debugging.

<!-- End of auto-generated description by mrge. -->