chore(scheduler): add groups table (#3986)
Adding a groups table in scheduler in order to 
- store the max concurrency value for each group 
- ensure each group is only processed by one processor at a time. (since
the algo is going to be 1. check how many slots are available for a
group 2. select candidate tasks for this group limit slots, we are gonna
have some locking at the group level)

    
<!-- This is an auto-generated description by mrge. -->
---

## Summary by mrge
Added a groups table to the scheduler to store max concurrency for each
group and ensure only one processor handles a group at a time.

<!-- End of auto-generated description by mrge. -->