fix(scheduler): allow skipping group update if locked
Adding an option to groups.upsert to allow skipping the update
if the group is being modified by another transaction.
When triggering action, the upsert is mostly used to set
the last_task_added_at. Because it is happens a lot in production this
query is waiting on other transaction. This commit allow to skip the
update since the last_task_added_at doesn't have to be 100% accurate.
It is currently only used to cleanup unused groups