chore(scheduler): decrease frequency of monitor worker
Expiring tasks isn't time critical. ie: it is ok if a task is not
considered expired for an extra second.
Since expiring tasks is the most expensive SQL query, we are lowering
its frequency to lower the pressure on the orchestrator db
This commit also add env vars for all the scheduler worker tick interval
so we can tweak the without having to make a code change