fix(orchestrator): fix dequeue long polling
I introduced a bug when adding dequeuing tasks by group key prefix:
The dequeue logic for early return would listen on `task:created:action*`
but the event is actually the real group key. ex: `task:created:action`
It led to the long polling mechanism of dequeue to never be notified
about created tasks while it was waiting.
The impact was very small in production because there is probably always
tasks to be ready to be dequeued and long polling is less useful there
but locally you would have had to wait for up to 10s (aka: next request)
for a task to be processed