fix(github-app-oauth): don't exit early on an update setup_action (#4027)
<!-- Describe the problem and your solution --> 
If a user is attempting to create a new connection using an Github app
that is already installed, allow the connection to be created.
Previously we just exited the flow once we saw it was an update action
for the `github-app-oauth`.

It is worth noting that updating the repositories an github-app-oauth
connects to for a user updates for all instances of that app for a user.
So if a user has connection `ABC` and connection `DEF` and updates the
repositories an app has access to for `ABC` it'll also update `DEF`

<!-- Issue ticket number and link (if applicable) -->

<!-- Testing instructions (skip if just adding/editing providers) -->

<!-- Summary by @propel-code-bot -->

---

This PR modifies the GitHub App OAuth flow to allow creation of new
connections using already installed GitHub apps by removing code that
was exiting early when detecting an 'update' setup action. The change
ensures users can create multiple connections using the same GitHub app
installation.

*This summary was automatically generated by @propel-code-bot*