feat(scheduler): add retry_key and owner_key to the tasks table
retry_key: when implementing a retry mechanism in the orchestrator I
forgot that we are gonna need to retrieve last attempts (or all attempts) of a given tasks.
Adding a retry_key column to task that will be shared by all retry
attempt. It will be used by async actions for example to be able to
retrieve the last attempt and return its output/error
owner_key: we are gonna need to ensure that retrieving a specific async
action comes from the actual environment that triggered it. We could
maintain a separate table to map a task and its environment but I prefer
to add an ownership feature to orchestrator. It will save an extra query
and it is more flexible and reusable.